Commit by Konstantin Kobel
Final commit of a one week practical course. During this week, the implementation of the mixed-model-ANOVA statistical test and the corresponding plots, including annotation of the statistical results within these plots was done. This final commit contains an update that allows more customization of the plots and the annotation features